refactor: merge core modules into single noco module
