+++ packages/nocodb/src/services/base-users/base-users.service.ts
@@ -350,7 +350,9 @@ export class BaseUsersService {
   // todo: refactor the whole function
   async sendInviteEmail(email: string, token: string, req: any): Promise<any> {
     try {
-      const template = (await import('./ui/emailTemplates/invite')).default;
+      const template = (
+        await import('~/services/base-users/ui/emailTemplates/invite')
+      ).default;
 
       const emailAdapter = await NcPluginMgrv2.emailAdapter();
