+++ packages/nocodb/src/services/datas.service.ts
@@ -4,8 +4,8 @@ import * as XLSX from 'xlsx';
 import papaparse from 'papaparse';
 import { nocoExecute } from 'nc-help';
 import type { BaseModelSqlv2 } from '~/db/BaseModelSqlv2';
-import type { PathParams } from '~/modules/datas/helpers';
-import { getDbRows, getViewAndModelByAliasOrId } from '~/modules/datas/helpers';
+import type { PathParams } from '~/helpers/dataHelpers';
+import { getDbRows, getViewAndModelByAliasOrId } from '~/helpers/dataHelpers';
 import { Base, Column, Model, Source, View } from '~/models';
 import { NcBaseError, NcError } from '~/helpers/catchError';
 import getAst from '~/helpers/getAst';
